把搬到其他地方的函数在原位置删干净
