+++ src/utils/shared.ts
@@ -225,3 +225,23 @@ export const formatTime = (date: Date, fmt: string) => {
   }
   return fmt;
 }
+
+export const changeTime = (x: number) => {
+  let res = ''
+  if (x >= 86400) {
+    res += Math.floor(x / 86400).toString() + ' 天 '
+    x -= Math.floor(x / 86400) * 86400
+  }
+  if (x >= 3600) {
+    res += Math.floor(x / 3600).toString() + ' 小时 '
+    x -= Math.floor(x / 3600) * 3600
+  }
+  if (x >= 60) {
+    res += Math.floor(x / 60).toString() + ' 分 '
+    x -= Math.floor(x / 60) * 60
+  }
+  if (x > 0) {
+    res += x.toString() + ' 秒 '
+  }
+  return res
+}
